Don't check for module, api and retpoline changes
I'm not sure why, but the build scripts are checking whether there are
any changes between this build and a previous kernel build. Having not
got any previous builds around to check, the build is failing for me
when it gets to this step.

I've not looked into why it's doing these checks, but should do - perhaps
I'll be able to use the build scripts in a different way that's more
suited to my use case. It feels odd that a build could depend on state left
behind from another build, so I'm not 100% sure that I've really
understood what's going on here.