Dynamically determine sources.list file for PPA
The path to the file containing the URL for System76's PPA repository on
launchpad has changed since I first wrote this script. I'm working the
path out at runtime (rather than hard coding the new path) so that it
continues to run on machines that were installed before System76 changed
the path, as they still use the old filenames.