Remove debootstrap chroot after building kernel
Note the call to `sleep` prior to unmounting `dev` and `proc`, which
have been bind-mounted inside the chroot. If the unmount commands fail
then we'll end up calling `rm -rf` on a directory tree that includes our
*actual* `/dev` and `/proc` directories. If you do that, you'll find
lots of things stop working. Restarting your system is the easiest fix.

So let's wait a decent amount of time (5 seconds) to give processes that
are using the `dev` or `proc` inside the chroot a good chance of
shutting down. This is required before we can unmount them before we run
`rm -rf`.