Build packages for latest available kernel
When I first wrote the script I wrote it so that it would download the
source for the currently installed kernel. Today when I ran it the
installed kernel was 5.11, but 5.13 was available in the repositories.
The `apt-get source` command downloaded 5.13, which broke the script as
it was expecting to work with `/usr/src/linux-5.11.0`.

This change means we determine the kernel version that we've downloaded
automatically by parsing the output of `apt-cache showsrc`.

I've also removed the option to specify the path to the chroot that
we'll use to build the kernel in. This feels sensible now that we're
forcefully removing the chroot after the build (see previous commit).