Install and hold custom kernel packages
Once you've built packages for a custom kernel you can install them in
the usual way, but on Pop!_OS there are several meta packages whose
dependencies conflict with your new packages.

At the time of writing the pop-desktop package depends on any version of
linux-system76, but that depends on the latest available versions of
the linux-image-generic and linux-headers-generic meta packages.

They, in turn, depend on specific versions of the packages that actually
contain the kernel, its modules and the header files.

This commit includes a script that builds some alternative meta packages
that can be installed in place of the Pop's standard ones. It uses the
`equivs-build` command from the equivs package. These meta packages
allow us to install our custom kernel packages without prompting `apt`
to re-install the default Pop! packages.

Use the `install-packages` script to install your new kernel and it will
install the meta packages at the same time. Then it will run `apt-mark`
to tell not `apt` not to upgrade any of them when you install updates
for the rest of your system.